Fix type conversion for currentTime in video event handling; change from parseInt to parseFloat for accurate time representation.

+++ src/content-script.ts
@@ -200,7 +200,7 @@ const bootstrap = () => {
           video.volume = Number.parseFloat(volumeValue)
           break
         case VIDEO_EVENTS.SEEKED: {
-          const time = Number.parseInt(currentTime)
+          const time = Number.parseFloat(currentTime)
           syntheticEvent = true
           video.currentTime = time
           break
